// src/oem/intermediaries/_oem-quotes-contacts/oem-quotes-contacts.enums/type.enum.ts
export enum TypeEnum {
  END_CUSTOMER = 'End Customer',
  PARTNER_SALES = 'Partner Sales',
  DISTRIBUTOR_SALES = 'Distributor Sales',
  DISTRIBUTOR_CONTACT = 'Distributor Contact',
  PARTNER_CONTACT = 'Partner Contact',
  TECHNICAL_CONTACT = 'Technical Contact',
  INTERNAL = 'Internal',
  OTHER = 'Other',
}